Make the detectron optional

+++ pipeline/pipeline.py
@@ -15,10 +15,22 @@ from preprocess import (
     append_background,
     blacklist,
 )
-from forecut import remove_bg
 from random import randint
 import sys
 from yolo_label_tools import count_from_top, find_pixel_edges, find_yolo_coordinates
+
+run_detectron = True
+try:
+    from forecut import remove_bg
+except:
+    detectron_input = input("""\nCould not import detectron2.\n
+Would you like to run the pipeline without removing backgrounds
+on nontransparent images? (y/n) """).upper()
+    if detectron_input != "Y":
+        print("Closing pipeline...")
+        sys.exit()
+    else:
+        run_detectron = False
 
 image_dir = "images"
 
@@ -154,25 +166,26 @@ for x in range(0, len(image_filepaths)):
 # exec(open("process_images.py").read())
 # {py process_images.py -i assets/images -p --cpus 2}
 
-for opaque_path in opaque_filepaths:  # e.g. opaque_path = images/tires/abc.jpg
-    remove_bg(opaque_path) # creates file without background
-    file_stem = os.path.splitext(opaque_path)[0] # images/tires/abc
-    output_file = f"output/{file_stem}.png"   # output/images/tires/abc.png
-    class_label = opaque_path.split('/')[-2] # tires
-    class_label_number = str(class_labels.index(class_label) + 1) # yolo counts from 1
-    coordinates = find_yolo_coordinates(output_file)
-    # remove blank images
-    if coordinates == None:
-        print('blank image:', output_file,"removing...")
-        os.remove(output_file)
-        os.remove(opaque_path)
-        continue
+if run_detectron:
+    for opaque_path in opaque_filepaths:  # e.g. opaque_path = images/tires/abc.jpg
+        remove_bg(opaque_path) # creates file without background
+        file_stem = os.path.splitext(opaque_path)[0] # images/tires/abc
+        output_file = f"output/{file_stem}.png"   # output/images/tires/abc.png
+        class_label = opaque_path.split('/')[-2] # tires
+        class_label_number = str(class_labels.index(class_label) + 1) # yolo counts from 1
+        coordinates = find_yolo_coordinates(output_file)
+        # remove blank images
+        if coordinates == None:
+            print('blank image:', output_file,"removing...")
+            os.remove(output_file)
+            os.remove(opaque_path)
+            continue
 
-    coordinates = [str(coordinate) for coordinate in coordinates]
-    line = ','.join([class_label_number] +  coordinates)
-    print('appending coordinates:',line)
-    with open(f'{file_stem}.txt','w') as f:
-        f.write(line)
+        coordinates = [str(coordinate) for coordinate in coordinates]
+        line = ','.join([class_label_number] +  coordinates)
+        print('appending coordinates:',line)
+        with open(f'{file_stem}.txt','w') as f:
+            f.write(line)
